Fix ItemController deleteto use long instead
fixes a bug where the delete item button would throw an exception

+++ Controllers/ItemController.cs
@@ -224,7 +224,7 @@ namespace CyberMall.Controllers
         [HttpPost, ActionName("Delete")]
         [Authorize]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(long id)
         {
             var item = await _context.ItemListings.FindAsync(id);
